\name{rOptManifold-package}
\alias{rOptManifold-package}
\alias{rOptManifold}
\docType{package}
\title{
This package optimizes function over variour matrix manifolds.
}
\description{

}
\details{
\tabular{ll}{
Package: \tab rOptManifold\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-13\cr
License: \tab What license is it under?\cr
}

}
\author{
Kejun He, Shiyuan He, Jianhua Huang, Yihui Xie

Maintainer: Kejun He<kejun@stat.tamu.edu>, Shiyuan He <shiyuan@stat.tamu.edu>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{

}
\examples{

}
